Form Error Messages and Flashing
Added messages to most or all of the form validators in forms.py.
Flashed messages now have categories determining what type of alert
(Bootstrap) they are. Validator errors are passed through a method in
routes.py to flash the danger alert. Other alerts are green or blue,
depending on the action.