Form persistence, cloning, and file fix
Forms on the editsplit page now fill in when you want to edit or clone
them. If you are not the user of a split and want to edit it, it makes a
new split for that user. Added a new line at the end of the split file,
as splitlines() was sometimes not reading the last line.